refactor: Replace bare `src` package with top-level actor_scrapy_books_example
Give the Actor a descriptively named top-level package instead of one
literally called `src`, run as `python -m actor_scrapy_books_example`.
Updates the Scrapy module references (settings SPIDER_MODULES/
NEWSPIDER_MODULE/ITEM_PIPELINES, SCRAPY_SETTINGS_MODULE, the items import),
scrapy.cfg, the Dockerfile, ruff/ty paths, and the README.
