+++ actor_scrapy_books_example/__main__.py
@@ -14,7 +14,7 @@ from apify.scrapy import initialize_logging, run_scrapy_actor
 from .main import main
 
 # Ensure the location to the Scrapy settings module is defined.
-os.environ['SCRAPY_SETTINGS_MODULE'] = 'src.settings'
+os.environ['SCRAPY_SETTINGS_MODULE'] = 'actor_scrapy_books_example.settings'
 
 
 if __name__ == '__main__':
