+++ scrapy.cfg
@@ -1,8 +1,8 @@
 [settings]
-default = src.settings
+default = actor_scrapy_books_example.settings
 
 [deploy]
-project = src
+project = actor_scrapy_books_example
 
 [apify]
-mainpy_location=src
+mainpy_location=actor_scrapy_books_example
